requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
Babel==2.13.0
bcrypt==4.0.1
blinker==1.6.2
Bootstrap-Flask==2.3.0
cachelib==0.9.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.3.2
defusedxml==0.7.1
exceptiongroup==1.1.3
Faker==19.6.2
Flask==2.3.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Dance==7.0.0
Flask-DebugToolbar==0.13.1
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-OpenID==1.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
Mako==1.2.4
MarkupSafe==2.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.5
URLObject==2.4.3
webassets==2.0
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
